data_ordering/requirements.txt
# Core dependencies
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0

# For perceptual hashing (Phase 5)
imagehash>=4.3.0

# For LLM integration (Phase 6)
google-generativeai>=0.3.0

# For loading .env files
python-dotenv>=1.0.0

# For GUI dialogs (Phase 4)
# Using tkinter which comes with Python
